+++ R/mod_metricGrid.R
@@ -52,7 +52,9 @@ metricGridServer <- function(id, metrics) {
             is_perc = metric['is_perc'] == 1,
             is_url = metric['is_url'] == 1,
             succ_icon = metric['succ_icon'],
-            icon_class = metric['icon_class'])
+            icon_class = metric['icon_class'],
+            type = metric['type']
+          )
         )
     })
   })