+++ R/mod_reportPreview.R
@@ -401,6 +401,7 @@ reportPreviewServer <- function(id, selected_pkg, maint_metrics, com_metrics,
       req(selected_pkg$name())
       
       tagList(
+        h5(code('{riskmetric}'), 'Assessment Date:'), selected_pkg$date_added(),
         if('Risk Score' %in% report_includes()) tagList(h5('Risk Score:'), selected_pkg$score()) else "",
         h5('Package Decision:'),ifelse(is.na(selected_pkg$decision()), 'Pending',selected_pkg$decision())
       )