+++ NEWS.md
@@ -1,3 +1,5 @@
+# riskassessment (development version)
+
 # riskassessment 1.0.0
 
 ### Enhancements