+++ NEWS.md
@@ -8,6 +8,7 @@
 * Introduced cards to Database tab, reviewing contents of db (#587)
 * Fixed busted button introduced with #547 (#592)
 * Allow `admin` users to edit roles and privileges (#541)
+* Prominently display the date a package was added (#486)
 
 
 # riskassessment 1.0.0