+++ NEWS.md
@@ -1,4 +1,4 @@
-# riskassessment (development version)
+# riskassessment 2.0.0
 * Disable using the Microsoft R Application Network (MRAN) for restoring packages
 * Removed legacy code that set the covr_coverage assessment's metric weight to 0 in the assessmentInfo mod and metric table initialization sql file.
 * Download source tarball when uploading package to database